Add support to build LLVM/clang with C and C++ support
We currently have a stale "llvm" branch, that does not build.
However, there is clear demand in the RISC-V toolchain community
for a working LLVM on top of a recent GNU toolchain.
In order to build such a toolchain, quite some LLVM and clang know-how is
required to avoid path issues at LLVM build time or later when using clang.
The main purpose of this commit is to demonstrate a way to combine
the RISC-V GNU toolchain repo with LLVM, with the intent to save
others hours of frustration, debugging time or support time.

patch from https://github.com/riscv-collab/riscv-gnu-toolchain/pull/1166/commits/425d91403c9a5ef36e991e1920de18ab6705c425

authored-by: Christoph Müllner <christoph.muellner@vrull.eu>